requirements.txt
Flask==2.3.3 
gunicorn==21.2.0 
pandas==2.1.4 
openpyxl==3.1.2 
playwright==1.40.0 
beautifulsoup4==4.12.2 
requests==2.31.0 
selenium==4.15.0 
phonenumbers==8.13.27 
python-dateutil==2.8.2 
PyYAML==6.0.1 
python-dotenv==1.0.0 
schedule==1.2.0 
redis==5.0.1 
pymysql==1.1.0 
click==8.1.7 
Werkzeug==3.0.1 
